Add migration for generic relations Tag model.
